requirements.txt
pyaudio
requests
RPi.GPIO
spidev
pyalsaaudio
